+++ generator/get.go
@@ -26,7 +26,7 @@ func writeGet(buf *bytes.Buffer, m *goparser.Method, stmt *sqlparser.Statement) 
 	for _, f := range stmt.Fields {
 		s := m.Results.Result()
 		v := s.VarByTag(f)
-		name := "xu." + v.Name()
+		name := "xu." + v.VName
 		w(v.Scan(name))
 		w(",")
 	}