+++ sqlparser/insert_test.go
@@ -9,8 +9,8 @@ import (
 
 func TestParseInsertStmt(t *testing.T) {
 	sql := `
-	insert into users(username, phone, address, status, birthday, created, updated)
-	values (${u.Username}, ${u.Phone}, ${u.Address}, ${u.Status}, ${u.Birthday},
+	insert into users_#{u.Username}(username, phone, address, status, birthday, created, updated)
+	values (?, ${u.Phone}, ${u.Address}, ${u.Status}, ${u.Birthday},
 	  CURRENT_TIMESTAMP, CURRENT_TIMESTAMP)`
 
 	p := NewParser(bytes.NewBufferString(sql))